# Clear overlay state vars in lsp-ui-doc--hide-frame