@kiennq 2022-05-18
I don't think that profiling is correct/applicable for this function since a lot of things are involved here (json-rpc perf, io queue pipe when the profiling is running, the profiling implementation details ...).
For example, using the same profiling function on my env provide these data:
```
move: Finished 1000 iterations in 11.041s (0.011s / iteration)
make: Finished 1000 iterations in 13.098s (0.013s / iteration)
master:  Finished 1000 iterations in 17.059s (0.017s / iteration)
```

So the version of using `overlay-buffer + move` is better than the version of `set to nil + make` a new overlay.

I think you want to benchmark the perf of making a new overlay compare to move overlay + `overlay-buffer` here.
A simple benchmark like this would be sufficient
``` el
(setq ov (make-overlay 1 10))
#<overlay from 1 to 10 in *scratch*>


(benchmark-run 100000 (progn
                        (delete-overlay ov)
                        (setq ov nil)
                        (setq ov (make-overlay 10 20))))
(0.061306730000000004 0 0.0)


(benchmark-run 100000 (progn
                        (delete-overlay ov)
                        (move-overlay ov 10 20)
                        (overlay-buffer ov)))
(0.034729481 0 0.0)
```
The `move + overlay-buffer` performs better.

Since overlay is being really deleted on garbage collection in Emacs, creating too much overlay can make the garbage collecting happens more frequently and sometimes slow down the whole Emacs.