@Lenbok 2022-05-18
Profiling the actual function is more useful than a toy example (as in your `progn`) as it demonstrates a more realistic usage pattern. I would like to also trigger redisplay to make it even more realistic - do you know how to do that? If it turns out the difference is negligible in the real-world scenario due to other factors dominating, then we are treading into premature optimization territory).

In case the profiler itself was causing problems as you suggested, I repeated my run using `benchmark-run` instead of the profiler and got essentially the same results: 
Master -> 18.7s
This PR -> 9.7s

What is your `move` version? The existing implementation on master supposedly already re-uses the overlay, does that mean it's actually broken?