@kiennq 2022-05-18
> What is your `move` version? The existing implementation on master supposedly already re-uses the overlay, does that mean it's actually broken?

I actually realize that I've changed in the wrong place by the' move' version so it would not take effect. So I retract the `move` version that I mentioned and use the master version instead.
Here is the benchmark version on the master (which use move overlay) and your PR.

``` elisp
(defun my-time-lsp-ui-doc-show()
  (interactive)
  (message "result: %s"
           (benchmark-run 1000
             (progn
               (lsp-ui-doc-show)
               (lsp-ui-doc-hide)))))
my-time-lsp-ui-doc-show

master: result: (9.237509808999999 4 0.6467752440000001)
pr: result: (11.377403877 5 0.7797696419999998)
```

Your PR which is resetting the overlay took more time and more garbage collection as well.

Boiling down the difference, it's mostly just the difference in using the `move-overlay` or `make-overlay` so my toy benchmark is actually working better here.

Simulating the "real" scenario as you said is being affected by a lot of external causes, for example, LSP perf might degrade the longer it runs, the same for Emacs memory.