@Lenbok 2022-05-18
Do you have `display-line-numbers-mode` enabled? I have been looking at why master is so slow for me compared to this PR, and that seems to be the difference. The overlay re-use seems incompatible with `display-line-numbers-mode`. Can you try to replicate?