@Lenbok 2022-05-18
BTW, my numbers using your benchmark function:

```
Current PR
Display-Line-Numbers mode enabled in current buffer
result: (9.8282039 9 1.2140667350000003)
Display-Line-Numbers mode disabled in current buffer
result: (10.308526724 9 1.5485936239999996)

Re-use overlay (master)
Display-Line-Numbers mode enabled in current buffer
result: (18.412163156000002 9 1.321076187)
Display-Line-Numbers mode disabled in current buffer
result: (9.649574025 9 1.2580851540000002)
```