@kiennq 2022-05-18
> Do you have `display-line-numbers-mode` enabled? I have been looking at why master is so slow for me compared to this PR, and that seems to be the difference.

Interesting. I do have `display-line-numbers-mode` enabled.
Which source file are you using to test?
I've been using this test file with cursor is placed in `<|>`
``` js
function fib(num) {
    if (num <= 1) {
        return num;
    }

    const result = fib(num - 1) + fib(num - 2);
    console.log(`fib(${num}): ${result}`);
    return result;
}

const f = fib(5);

const cp = require("child_process");
cp.s<|>pawn("./a.cmd");

class Foo {
    size = 4;
    soo() { }
    ["so so"]() { }
}

let a = new Foo();
```