@Lenbok 2022-05-18
I don't have javascript set up, I did do one of the server installs and there are no real docs for `spawn` (I assume that needs setting up some additional libraries). With the cursor on `require` there is a doc of a couple of lines. With that there are no significant differences when running the benchmarking on master with/without line numbers.

I suspect that since it's related to `display-line-numbers-mode` it may depend on the length of the documentation and/or the source file and how they relate to the overlay. For the previous benchmarking I've been using openscad (a file about 80 lines long, and a function doc about 10 lines long) and I also tried some python that was slower overall and when `display-line-numbers-mode` is active it is slower but by only a few seconds. Both of those were bringing up docs from other source files, so not so easy to share. I did an experiment of computing the inline window width once instead of calling `lsp-ui-doc--inline-window-width` for every line of the doc and that made most of the difference go away:

```
modified   lsp-ui-doc.el
@@ -745,8 +745,9 @@ FN is the function to call on click."
     string))
 
 (defvar-local lsp-ui-doc--inline-width nil)
+(defvar-local lsp-ui-doc--inline-window-width nil)
 
-(defun lsp-ui-doc--inline-window-width nil
+(defun lsp-ui-doc--calc-inline-window-width nil
   (- (min (window-text-width) (window-body-width))
      (if (bound-and-true-p display-line-numbers-mode)
          (+ 2 (line-number-display-width))
@@ -754,11 +755,10 @@ FN is the function to call on click."
      1))
 
 (defun lsp-ui-doc--inline-zip (s1 s2)
-  (let* ((width (lsp-ui-doc--inline-window-width))
-         (max-s1 (- width lsp-ui-doc--inline-width 2)))
+  (let* ((max-s1 (- lsp-ui-doc--inline-window-width lsp-ui-doc--inline-width 2)))
     (truncate-string-to-width
      (concat (truncate-string-to-width s1 max-s1 nil ?\s) s2)
-     width nil ?\s)))
+     lsp-ui-doc--inline-window-width nil ?\s)))
 
 (defun lsp-ui-doc--inline-padding (string len)
   (let ((string (concat " " string (make-string (- len (string-width string)) ?\s) " ")))
@@ -768,6 +768,7 @@ FN is the function to call on click."
 (defun lsp-ui-doc--inline-faking-frame (doc-strings)
   (let* ((len-max (-max-by '> (-map 'string-width doc-strings))))
     (setq lsp-ui-doc--inline-width len-max)
+    (setq lsp-ui-doc--inline-window-width (lsp-ui-doc--calc-inline-window-width))
     (--map (lsp-ui-doc--inline-padding it len-max) doc-strings)))
 
 (defun lsp-ui-doc--inline-untab (string)

```