@Lenbok 2022-05-20
There are too many other things broken with lsp-ui-doc, it needs wider changes. Closed in favor of #711